Moves and renames fortls client to lsp dir
